feat: sample data + test page
